Create YHOW.java
This is an Agent-Based Simulation. This is for simulating the Muscle wilt in males starting at age 50. It is run on a 10-year run. This simulation shows the progression of the dead cells infecting the healthy ones. If the patient is injected with a Youth vaccine that prevents Muscle Wilt, then my simulation also simulates the dead cells returning to healthy ones.